+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [1.3.0] - 2018-10-09
+### Added
+- new option "--no-cache-clean"
+- new option "--no-cache-flush"
+- new option "--no-progress"
+- new option "--no-clean-url-key"
+
 ### Changed
 - optimized code
 - modified logic of url rewrites regeneration